Add core-component-page when pushing to gh-pages
Currently the core-component-page is not loaded in components
and thus page gets a "Failed to fetch resource" error. Adding it
to bower is not an option as it would polute the dependencies.
Cloning the entire repository wasn't a good idea either because
it would have to be registerd as a git submodule -
https://help.github.com/articles/page-build-failed-missing-submodule