fix vulcanize to bundle absolute url paths in the entrypoint (#83)
* fix vulcanize to bundle absolute url paths in the entrypoint

* fix bad root in path, root would never be outside of cwd

* better fix for vulcanize: use proper target url instead of forcing abspath

I've updated this to use a better fix since we don't want to use abspath here. Instead, we're fixing the hydrolosis->polymer-analyzer compatibility loader to handle the absolute paths vulcanize can send it. This lets us pass an absolute file path in inputUrl, which is what vulcanize was actually expecting. Sending relative paths through this option is what was previously breaking some projects.