Improve ES6 module support. (#428)
This ensures that we don't throw parse errors for ES6 modules. We can't, in general, know at parse time whether a given bit of JS should be parsed as a module or a script.

I think, given that, it is better to parse robustly and just note whether a file must be a module, or that it is legal syntax as either a module or a script.