Check Edge versions correctly in compilation switch. (#174)
Really fixes #161. This time for real! My previous PR #173 assumed that
the Edge user agent used the main browser release version. Actually it
uses the EdgeHTML version. The fix we care about for the original issue
is in a major release of Edge, but a minor release of EdgeHTML, so we
have to check minor version now too.

https://en.wikipedia.org/wiki/Microsoft_Edge#Release_history

Also added a test and a small refactor so that we can paste real user
agent strings.

Also note this uses destructuring, so we are no longer node 4 compatible.