Exclude neon-animation/web-animations.html
This file should only be referenced by HTML files, and with this commit it will remain as an HTML Import. This isn't entirely ideal, as it would be better if it were rewritten as a script tag. Filed as https://github.com/Polymer/polymer-modulizer/issues/162 to follow up with a cleaner fix.

Also renamed ConverterMetadata to ConversionMetadata.

Fixes https://github.com/Polymer/polymer-modulizer/issues/152