Split Closure type expression parsing from TypeScript serialization.
This gives us an intermediate AST for type expressions that we can walk
and transform. In particular, this will be helpful for an upcoming PR
where we will try to resolve all names in type expressions (e.g. to add
triple-slash references, and to warn/replace with `any` when unresolvable).