Handle method rest parameters.
Before this change, rest parameters (e.g. `foo(...args)`) would have the
`...` included in their name, and would not include any information from
their JSDoc annotation.

Now the parameter name does not include the `...`, and the JSDoc
annotations are correctly linked (note that the type includes the `...`
to make it clear this is a rest parameter). This is how regular
functions already work in Analyzer.