Only create <reference> statements for outbound dependencies.
Sometimes an Analyzer document includes an import feature that is
inbound (things that depend on me) instead of outbound (things I
depend on). For example, if an HTML file has a <script> tag for a JS
file, then the JS file's Analyzer document will include that <script>
tag as an import feature. We only care about outbound dependencies.