Fix import paths that are now resolved as of Analyzer 2.5.0.
When we analyze a package's Git repo, our dependencies are installed to
"<repo>/bower_components". However, when this package is itself installed
as a dependency, our own dependencies will instead be siblings, one
directory up the tree.

Analyzer (since 2.5.0) will set an import feature's URL to the resolved
dependency path as discovered on disk. An import for "../foo/foo.html"
will be resolved to "bower_components/foo/foo.html". Transform the URL
back to the style that will work when this package is installed as a
dependency.