fix silent (!) build failure when several source filepaths cannot be found (#276)
* add a (failing) test case for https://github.com/Polymer/polymer-cli/issues/936

* ensure the build fails when more than one source file is missing

The 'analyze' phase of the build process completes when all source files are analyzed (scanned). A source file is analized when all its dependencies are resolved or rejected. When the sources stream is complete, _all_ pending source file promises need to be rejected--not only the first one. If some of the deferred dependencies remain pending, the analysis of the toplevel source file never continues after 'await this._cache.dependencyGraph.whenReady(resolvedUrl);'. That leads of a premature end of the analysis phase (without any errors).

* update CHANGELOG

* add test for lazy-import with broken path (#1)

add a test case to verify that this PR also fixes the issue https://github.com/Polymer/polymer-cli/issues/904