Simplfy GenericOptimizeTransform interface.
No longer takes an `any` options to pass to transform function. Instead, just
assumes the subclass will handle options itself.

This is motivated by an upcoming change, where the JSCompileTransform is
going to need access to the File object.

Got rid of JSBabelTransform class because it was now doing almost
nothing.