Add common tsconfig-base.json and connect to one package.
The idea is to have a strict base config in the top-level of tools/.
Each package extends the base, sets relative source paths (which are
always interpreted relative to the file that references them), and any
required additional compiler flags.

The "extends" syntax requires gulp-typescript >= 3, so I did that
upgrade in Analyzer.

Will add other projects if we like this idea.