Don't AMD transform non-module scripts.
We now pay more attention to whether an HTML script has type=module via
the HTML splitter, and only run the AMD transform on those that do.

For external scripts, we now only run the AMD transform if the file
looks like a module based on whether it has import or export statements.
Babel 7 can do this for us with the "unambiguous" sourceType.