Preserve comments more reliably.
Sometimes we were moving comments out of HTML and onto the first JS node. This is fine, except that we were then pruning that first JS node!

We should probably move to always preserving comments when pruning/replacing nodes, but this also gives us a place to let people know where these comments came from, and that they need review.