`Feature#astNode` is now an AstNodeWithLanguage
This removes the largest remaining use of `any` in the analyzer's API. What previously was an AST node is now wrapped with a union discriminated by the language name. This way, a user can process an ast node that might be a parse5 node, or a babel node, or even a future language, all in a type safe way.

The main place where we'll get errors without Typescript warning is is where we're doing `babel.isSpecificTypeOfNode(feature.astNode)`. The babel-types API is very permissive in the types of the values it accepts, meaning that this will not warn, but it will now always return false. We need to change that to `feature.astNode.node`.