Apply the `no-any` lint rule to all repos.
This will warn if we write the type `any` anywhere. This is bad because
`any` is viral, and silently turns off all type checking anywhere it is
used.

Also got all repos on the latest version of tslint.

Two tricks that came up a lot here:
  - The type `null | undefined | {}` accepts any value, but almost
    nothing is  legal to do to that type. It's a better way of
    expressing "I don't touch this value without without verifying
    my assumptions first."
  - The related type `Partial<Error> | string | null | undefined` is a
    pretty good representation of our expectations around what an error
    might be.

I only fixed the really quick and simple cases. There are still a number
of places where we're still using `any` in scary ways.