Break out more shop tests.
Adds tests of the source with a few different polyserve definitions.

Also breaks out each of the individual build configuration tests.

This isn't bad performance-wise, because the vast majority of the cost is in `npm install` and `polymer build` which is not repeated.

Also removes some unnecessary and distracting logging statements.