Collapse ephemeral super classes into their children.
If a class is annotated as @private, and is used as the superclass of
one or more other classes, then it will be merged into its child
classes, and not itself emitted as a feature.

This is to handle the case where a superclass is declared as a separate
variable, usually so that mixins can be applied in a way that is
compatible with the Closure compiler.