Modifies the integration test golden format to handle test page failures. (...)
If a test run completely fails to load the test page, then the results returned
by `BrowserRunner` do not include counts for passing, pending, and failed
tests. This commit modifies the format of the golden to nest these 'stats' into
their own distinct object, which allows the entire object to be tested for deep
equality because Mocha's object deep equality test (correctly) considers lack
of a property and an property being set to `undefined` to be unequal. (i.e.
`{}` does not 'deeply equal' `{p: undefined}`.)